Use path as the cache key for getTypeExports
On large monorepos, hashing the file contents was causing this plugin
to consume the majority of the time in any eslint run. I can see
hashing the file content being useful if you have multiple identical
copies of the same file in different places in the repo, but that should
be an extreme corner-case.